refactor: step2 피드백 적용 - 객체 책임 위임
