+++ src/main/java/nextstep/subway/line/domain/Lines.java
@@ -18,33 +18,9 @@ public class Lines {
         WeightedMultigraph<Station, DefaultWeightedEdge> graph = new WeightedMultigraph(DefaultWeightedEdge.class);
 
         for (Line line : lines) {
-            registerPath(line, graph);
+            line.registerPath(graph);
         }
 
         return new DijkstraShortestPath(graph);
-    }
-
-    private void registerPath(Line line, WeightedMultigraph<Station, DefaultWeightedEdge> graph) {
-        addVertices(line, graph);
-        addEdges(line, graph);
-    }
-
-    private void addVertices(Line line, WeightedMultigraph<Station, DefaultWeightedEdge> graph) {
-        line.findStations()
-                .forEach(graph::addVertex);
-    }
-
-    private void addEdges(Line line, WeightedMultigraph<Station, DefaultWeightedEdge> graph) {
-        line.getSections()
-                .getSections()
-                .forEach(section -> addEdgeWith(addEdge(section, graph), section.getDistance(), graph));
-    }
-
-    private DefaultWeightedEdge addEdge(Section section, WeightedMultigraph<Station, DefaultWeightedEdge> graph) {
-        return graph.addEdge(section.getUpStation(), section.getDownStation());
-    }
-
-    private void addEdgeWith(DefaultWeightedEdge edge, int weight, WeightedMultigraph<Station, DefaultWeightedEdge> graph) {
-        graph.setEdgeWeight(edge, weight);
     }
 }
